requirements.txt
asn1crypto==0.24.0
aws==0.2.5
awscli==1.16.1
bcrypt==3.1.4
boto==2.49.0
boto3==1.8.1
botocore==1.11.1
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
Click==7.0
colorama==0.3.9
configparser==3.7.1
cryptography==2.3.1
docutils==0.14
fabric==2.3.1
Flask==1.0.2
idna==2.7
invoke==1.1.1
itsdangerous==1.1.0
Jinja2==2.10
jmespath==0.9.3
MarkupSafe==1.1.0
paramiko==2.4.1
pdfkit==0.6.1
Pillow==5.4.1
prettytable==0.7.2
pyasn1==0.4.4
pycparser==2.18
PyNaCl==1.2.1
python-dateutil==2.7.3
PyYAML==3.13
requests==2.21.0
rsa==3.4.2
s3transfer==0.1.13
six==1.11.0
urllib3==1.23
virtualenv==16.0.0
Werkzeug==0.14.1
